+++ toolkit/scaffold/project/cpp/qt5/example.py
@@ -16,7 +16,7 @@ def create_example():
     pass
 
 
-class ExampleQmlContext(BaseModel):
+class ExampleContext(BaseModel):
     qt_version: str = "5.12.6"
     qt_tool_chain: list = [
         ["mingw73_64", "mingw730_64"],
@@ -30,25 +30,25 @@ class ExampleQmlContext(BaseModel):
     qt_tool_version: str = None
 
 
-def example_qml_user_input_context_hook(context: dict) -> dict:
-    example_qml_context = ExampleQmlContext(x64_arch=False)
+def example_user_input_context_hook(context: dict) -> dict:
+    example_context = ExampleContext(x64_arch=False)
 
-    example_qml_user_input_context = ExampleQmlContext.parse_obj(context)
+    example_user_input_context = ExampleContext.parse_obj(context)
 
     (
-        example_qml_context.qt_compile_version,
-        example_qml_context.qt_tool_version,
-    ) = example_qml_user_input_context.qt_tool_chain
+        example_context.qt_compile_version,
+        example_context.qt_tool_version,
+    ) = example_user_input_context.qt_tool_chain
 
-    if "64" in example_qml_context.qt_compile_version:
-        example_qml_context.x64_arch = True
+    if "64" in example_context.qt_compile_version:
+        example_context.x64_arch = True
 
-    log.debug(f"example_qml_context: {example_qml_context}")
+    log.debug(f"example_context: {example_context}")
 
     return (
         context
-        | example_qml_context.dict(exclude_none=True)
-        | example_qml_user_input_context.dict(exclude_none=True)
+        | example_context.dict(exclude_none=True)
+        | example_user_input_context.dict(exclude_none=True)
     )
 
 
@@ -56,8 +56,18 @@ create_example_qml = generate_create_project_command(
     command_help="Create a cpp qt5 qml example project scaffold.",
     template_paths=TEMPLATE_CPP_QT5_EXAMPLE_PATH / "qml",
     raw_user_input_context=USER_INPUT_CONTEXT
-    | ExampleQmlContext().dict(exclude_none=True),
-    user_input_context_hook=example_qml_user_input_context_hook,
+    | ExampleContext().dict(exclude_none=True),
+    user_input_context_hook=example_user_input_context_hook,
 )
 
 create_example.add_command(create_example_qml, "qml")
+
+create_example_console = generate_create_project_command(
+    command_help="Create a cpp qt5 console example project scaffold.",
+    template_paths=TEMPLATE_CPP_QT5_EXAMPLE_PATH / "console",
+    raw_user_input_context=USER_INPUT_CONTEXT
+    | ExampleContext().dict(exclude_none=True),
+    user_input_context_hook=example_user_input_context_hook,
+)
+
+create_example.add_command(create_example_console, "console")